+++ packages/codebooth/CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.3.0 (March 26, 2023)
 
-- add `scrollBehavior` option to `<Replay>` and `<ReplayMultiple>`
+- add `scrollBehavior` and `shouldScroll` options to `<Replay>` and `<ReplayMultiple>`